Separate out get_random_weight_neighbour
